flask_neo_pixel/requirements.txt
Adafruit-Blinka==6.4.1
adafruit-circuitpython-neopixel==6.0.2
adafruit-circuitpython-pypixelbuf==2.2.4
Adafruit-PlatformDetect==3.4.0
Adafruit-PureIO==1.1.8
asn1crypto==0.24.0
attrs==20.3.0
certifi==2018.8.24
chardet==3.0.4
click==7.1.2
cryptography==2.6.1
entrypoints==0.3
Flask==1.1.2
flask-expects-json==1.5.0
idna==2.6
importlib-metadata==3.7.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
keyring==17.1.1
keyrings.alt==3.1.1
MarkupSafe==1.1.1
pycrypto==2.6.1
pyftdi==0.52.9
PyGObject==3.30.4
pyrsistent==0.17.3
pyserial==3.5
python-apt==1.8.4.3
pyusb==1.1.1
pyxdg==0.25
requests==2.21.0
rpi-ws281x==4.2.6
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
sysv-ipc==1.1.0
typing-extensions==3.7.4.3
urllib3==1.24.1
Werkzeug==1.0.1
zipp==3.4.1
